init use only one sigv4 file
